+++ kedro_kubeflow/hooks.py
@@ -1,6 +1,8 @@
 import os
-from typing import Any, Dict
+from os import getenv
+from typing import Any, Dict, Iterable
 
+from kedro.config import ConfigLoader, TemplatedConfigLoader
 from kedro.framework.context import load_context
 from kedro.framework.hooks import hook_impl
 from kedro.io import DataCatalog
@@ -35,4 +37,23 @@ class MLFlowActivateParentHook:
         mlflow.start_run(run_id=os.getenv("MLFLOW_PARENT_ID"))
 
 
+class RegisterTemplatedConfigLoaderHook:
+    """Provides config loader that can substitute $(commit_id) and $(branch_name)
+    placeholders with information taken from env variables."""
+
+    @hook_impl
+    def register_config_loader(
+        self, conf_paths: Iterable[str]
+    ) -> ConfigLoader:
+        return TemplatedConfigLoader(
+            conf_paths,
+            globals_dict={
+                "commit_id": getenv("KEDRO_KUBEFLOW_COMMIT", default=None),
+                "branch_name": getenv("KEDRO_KUBEFLOW_BRANCH", default=None),
+            },
+        )
+
+
 mlflow_activate_parent_run = MLFlowActivateParentHook()
+
+register_templated_config_loader = RegisterTemplatedConfigLoaderHook()