+++ tests/test_function.py
@@ -1,5 +1,0 @@
-"""Test kedro_kubeflow functions."""
-
-
-def test_app():
-    pass